Proof of concept integration of the Android-5 Account/Credentials storage
This is a very quick and dirty hack rather than a proper implementation as it's only implemented for IMAP account and only used when the user Validate the "Incoming server" setting page of an already existing IMAP account.

I also hit a bug (or so I think) where AccountManager.updateCredentials() throws an exception when authTokenType is null although the API doc states it's allowed. Consequently I had to comment out some code of the underlying authenticator code to ignore that argument when updating credentials.